fix: Use --no-default-features
If --no-default-features is NOT used, the the selected features used in
a given build are those selected by cargo-feature-matrix PLUS the
default features.   This is contrary to the UI output and likely
contrary to the desired behavior of a test matrix.

This commit adds --no-default-features to the command arguments.  A new
test case is also added which will fail if the default features are not
properly supressed.

Along the way, the `features` sample was updated to use a new version of
`rand`, as the version specified doesn't actually compile with a current
stable compiler.